// src/app/interfaces/pokemons-response.ts
// Generated by https://quicktype.io

export interface PokemonsReponse {
    count:    number;
    next:     string;
    previous: null;
    results:  Result[];
  }
  
  export interface Result {
    name: string;
    url:  string;
  }

  export interface PokemonType {
    name: string;
    url: string;
  }